all tests working now
