+++ src/main/java/com/metalr2/config/constants/ViewNames.java
@@ -20,7 +20,6 @@ public class ViewNames {
   public static class Frontend {
     public static final String SEARCH_ARTISTS            = "frontend/search-artists";
     public static final String ARTIST_DETAILS            = "frontend/artist-details";
-    public static final String ARTIST_SEARCH_RESULTS     = "frontend/artist-search-results";
     public static final String SETTINGS                  = "frontend/settings";
     public static final String PROFILE                   = "frontend/profile";
     public static final String ARTISTS_RELEASES          = "frontend/artists-releases";
