+++ src/main/resources/templates/frontend/fragments/layout.html
@@ -18,6 +18,8 @@
   <script th:src="@{/webjars/popper.js/umd/popper.min.js}" src="#"></script>
   <script th:src="@{/webjars/bootstrap/js/bootstrap.min.js}" src="#"></script>
   <script th:src="@{/js/common.js}" src="#"></script>
+  <script th:src="@{/js/search-artists.js}" src="#"></script>
+  <script th:src="@{/js/follow-artists.js}" src="#"></script>
 </head>
   <body>
     <header>
